refactor: extract Role enum to separate file to fix circular import
