+++ src/models/__init__.py
@@ -1,6 +1,8 @@
-from enum import Enum
+from .role import Role
+from .person import Person
+from .defense import Defense
+from .room import Room
+from .time_slot import TimeSlot
+from .session_parameters import SessionParameters
 
-class Role(Enum):
-    SUPERVISOR = "supervisor"
-    REVIEWER = "reviewer"
-    CHAIRMAN = "chairman"
+__all__ = ['Role', 'Person', 'Defense', 'Room', 'TimeSlot', 'SessionParameters']
