+++ src/models/role.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class Role(Enum):
+    SUPERVISOR = "supervisor"
+    REVIEWER = "reviewer"
+    CHAIRMAN = "chairman"
